# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spark Lead Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /auth/set-password:
    post:
      operationId: setPassword
      tags: [auth]
      summary: Set password via token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordRequest"
      responses:
        "200":
          description: Password set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /companies:
    get:
      operationId: getCompanies
      tags: [companies]
      summary: Get companies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Companies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"

    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyInput"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyInput"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /services/{id}/companies:
    get:
      operationId: getServiceCompanies
      tags: [services]
      summary: Get companies linked to a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"

    put:
      operationId: linkServiceCompanies
      tags: [services]
      summary: Update company links for a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkCompaniesInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /analytics/stats:
    get:
      operationId: getAnalyticsStats
      tags: [analytics]
      summary: Get analytics stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsStats"
  /analytics/lead-trend:
    get:
      operationId: getLeadTrend
      tags: [analytics]
      summary: Lead creation trend last 30 days
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
  /analytics/stage-distribution:
    get:
      operationId: getStageDistribution
      tags: [analytics]
      summary: Lead count per stage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Distribution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StageCount"
  /analytics/kill-reasons:
    get:
      operationId: getKillReasons
      tags: [analytics]
      summary: Kill reason counts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Kill reasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KillReasonCount"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
      required:
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        permissions:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: boolean
      required:
        - token
        - user
        - permissions
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
        isWhitelisted:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - displayName
        - role
        - isWhitelisted
        - createdAt
    SetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required:
        - token
        - password
...
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdBy:
          type: string
          nullable: true
        createdAt:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceBrief"
      required:
        - id
        - name
        - createdAt
    CreateCompanyInput:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
    UpdateCompanyInput:
      type: object
      properties:
        name:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    ServiceBrief:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
...
    LinkCompaniesInput:
      type: object
      properties:
        companyIds:
          type: array
          items:
            type: string
      required:
        - companyIds
...
    AnalyticsStats:
      type: object
      properties:
        avgConversionDays:
          type: number
          nullable: true
        winRate:
          type: number
        lostCount:
          type: integer
        activePipelineCount:
          type: integer
      required:
        - winRate
        - lostCount
        - activePipelineCount
    TrendPoint:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required:
        - date
        - count
    StageCount:
      type: object
      properties:
        stage:
          type: string
        count:
          type: integer
      required:
        - stage
        - count
    KillReasonCount:
      type: object
      properties:
        reason:
          type: string
        count:
          type: integer
      required:
        - reason
        - count
...